nepse_scraper/api_schema.py
from drf_yasg import openapi
from drf_yasg.views import get_schema_view
from rest_framework import permissions

schema_view = get_schema_view(
    openapi.Info(
        title="NEPSE Data Scraper API",
        default_version='v1',
        description="""
        API for scraping and retrieving Nepal Stock Exchange (NEPSE) data.
        """,
        terms_of_service="https://www.google.com/policies/terms/",
        contact=openapi.Contact(email="contact@nepse.local"),
        license=openapi.License(name="BSD License"),
    ),
    public=True,
    permission_classes=[permissions.AllowAny],
)