requirements.txt
# Core
Django==4.2.7
djangorestframework==3.14.0
drf-yasg==1.21.7

# Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3